Remove mysql adapter
