--- specs/adapters/mysql_adapter_spec.rb
@@ -1,15 +1,0 @@
-require 'spec_helper'
-
-describe "Spectacles::SchemaStatements::MysqlAdapter" do
-  config = {
-    :adapter => "mysql",
-    :host => "localhost",
-    :username => "root"
-  }
-  configure_database(config)
-  recreate_database("spectacles_test")
-  load_schema
-
-  it_behaves_like "an adapter", "MysqlAdapter"
-  it_behaves_like "a view model"
-end
